% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{as_persistence}
\alias{as_persistence}
\alias{persistence}
\alias{diagram}
\alias{persistence_diagram}
\alias{as_persistence.default}
\alias{as_persistence.persistence}
\alias{as_persistence.list}
\alias{as_persistence.diagram}
\alias{as_persistence.PHom}
\alias{print.persistence}
\alias{format.persistence}
\alias{get_pairs}
\alias{as.data.frame.persistence}
\title{Efficient Format for Persistence Data (Diagrams)}
\usage{
as_persistence(x, ...)

\method{as_persistence}{default}(x, modulus = NULL, max_dim = NULL, threshold = NULL, ...)

\method{as_persistence}{persistence}(x, ...)

\method{as_persistence}{list}(x, ...)

\method{as_persistence}{diagram}(x, ...)

\method{as_persistence}{PHom}(x, ...)

\method{print}{persistence}(x, ...)

\method{format}{persistence}(x, ...)

get_pairs(x, degree, ...)

\method{as.data.frame}{persistence}(x, row.names = NULL, optional = TRUE, ...)
}
\arguments{
\item{x}{Persistence data to be transformed; either a ≥3-column matrix (or
object coercible to one) with dimension/degree, start/birth, and end/death
columns, a list whose first element is such an object, an object of class
'PHom' as returned by \verb{[ripserr::vietoris_rips()]}, or (a list as returned
by a \verb{*Diag()} function in \strong{TDA} (e.g. \verb{[TDA::ripsDiag()]}) whose first
element is) an object of class 'diagram'.}

\item{...}{Parameters passed to methods.}

\item{modulus, max_dim, threshold}{Possibly missing parameters of the
calculation of the persistence data \code{x} to be included in the 'persistence'
object.}

\item{degree}{Non-negative integer; the homology degree for which to recover
a matrix of persistence pairs.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}
}
\value{
Persistence data in the form of a list with named entries:
\item{\code{pairs}}{
A list of 2-column matrices containing birth-death pairs
in each dimension/degree, starting with zero.
}
\item{\code{modulus}}{
The (integer) modulus of the prime field in which
homology coefficientts were calculated.
}
\item{\code{max_dim}}{
The (integer) maximum dimension of calculated features.
}
\item{\code{threshold}}{
The (real) maximum length of an edge in the simplicial filtration.
}
}
\description{
Transform persistence data to the format used by \strong{plt}.
}
\details{
This class and function will soon be spun off to a lower-level package that
can be imported by others as needed. Refer to
\url{http://adv-r.had.co.nz/S3.html} for guidance on methods to make available.
}
\examples{

# sample from a cylinder
x <- tdaunif::sample_cylinder_elliptical(n = 60)
# calculate PH using {ripserr}
ph <- ripserr::vietoris_rips(x, dim = 1)
as_persistence(ph)
# calculate PH using {TDA}
ph <- TDA::alphaComplexDiag(x, maxdimension = 1)
as_persistence(ph)
# create a 'persistence' object directly from a matrix
as_persistence(unclass(ph$diagram))
}
