% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_code.R
\name{eval_student_code}
\alias{eval_student_code}
\alias{eval_answer_code}
\alias{eval_code_secure}
\title{Secured Code Evaluation}
\usage{
eval_student_code(file_name, expr, user_id = "ag", fixed_path = FALSE)

eval_answer_code(file_name, expr, user_id = "root", fixed_path = FALSE)

eval_code_secure(file_path, expr, user_id = unix::getuid())
}
\arguments{
\item{file_name, file_path}{Name of the file or path to the file location.}

\item{expr}{Code to evaluate.}

\item{user_id}{Name of the user to run the code under}

\item{fixed_path}{Ensure an absolute or relative path to file.}
}
\value{
Result of executed code.
}
\description{
Process the code using a different user outside of the root configuration.
}
\details{
Understanding the function wrappers:
\itemize{
\item \code{eval_code_secure()} is the workhorse engine.
\item \code{eval_student_code()} provides an interface to run \strong{student code}
relative to where it resides on the path.
\item \code{eval_answer_code()} provides an interface to run \strong{answer key} code
relative to where it resides on the path.
}
}
\section{Code Evaluation}{


Place any code that should be evaluated within curly brackets, e.g. \code{{}}.
}

\examples{

\dontrun{
# Obtain student results
eval_student_code("student_code.R", { cat("hello!") })
}

\dontrun{
# Obtain solution results
eval_student_code("student_code.R", { cat("hello!") })
}

\dontrun{
# Generalize it to work with any user.
eval_student_code("student_code.R", { cat("hello!") }, user_id = "anyone")
}
}
