% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-format-to-pl.R
\name{format_testthat_to_pl}
\alias{format_testthat_to_pl}
\title{Construct the PrairieLearn output format JSON}
\usage{
format_testthat_to_pl(test_data)
}
\arguments{
\item{test_data}{A \code{data.frame} containing the test information.}
}
\value{
Formats the \code{data.frame} into a \code{list} for export into JSON.
}
\description{
Interprets and builds the appropriate message for each test result in
the PrairieLearn format.
}
