% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-helpers.R
\name{path_student}
\alias{path_student}
\alias{path_answer}
\title{Path helpers}
\usage{
path_student(file_name, fixed_path = FALSE)

path_answer(file_name, fixed_path = FALSE)
}
\arguments{
\item{file_name}{Name of the file to construct the file path.}

\item{fixed_path}{Create an absolute (e.g. root \verb{/} path) or relative.
Default is a relative path.}
}
\description{
Expand path to code file location on a PL grader container.
}
