% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,Rcpp_PersistenceLandscape-method}
\title{Plot Method for Persistence Landscapes.}
\usage{
\S4method{plot}{Rcpp_PersistenceLandscape}(
  x,
  replace_inf = NULL,
  n_levels = NULL,
  palette = "viridis",
  alpha = NULL,
  rev = FALSE,
  ...,
  silent = TRUE
)
}
\arguments{
\item{x}{A persistence landscape object of class 'Rcpp_PersistenceLandscape'.}

\item{replace_inf}{When using an exact representation of a persistence
landscape, infinite values can appear. If not \code{NULL}, this value will
replace \code{Inf} (and its negative will replace \code{-Inf}) in the plot.}

\item{n_levels}{Integer; number of levels to plot. If \code{NULL} (the default),
determined to be the number of levels in \code{pl} or \code{x}.}

\item{palette}{Character; either a color palette from
\code{\link[grDevices:palettes]{grDevices::hcl.pals()}}, or the name of a palette function documented there
(e.g. \code{'rainbow()'}), or a vector of colors for \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}} to
interpolate.}

\item{alpha, rev}{Parameters passed to \code{\link[grDevices:palettes]{grDevices::hcl.colors()}},
\code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}

\item{...}{Additional parameters passed to \code{\link[base:plot]{base::plot()}}. Values passed to
\code{type} or \code{col} will be ignored with a message.}

\item{silent}{Logical; whether to silence messages.}
}
\description{
A \code{\link[base:plot]{base::plot()}} method for persistence landscape objects.
}
\examples{
# sample points
set.seed(650637L)
rpp <- tdaunif::sample_projective_plane(48L)

# compute persistence data, retaining parameters
pd <- as_persistence(ripserr::vietoris_rips(rpp, max_dim = 2L, threshold = 4))

# plot landscapes
par(mfrow = c(3L, 1L), mar = c(0, 2, 0, 2))
# palette name
plot(landscape(pd, degree = 0L, xmin = 0, xmax = 1.5, by = 0.01),
     palette = "Cork", lwd = 1)
# palette function name
plot(landscape(pd, degree = 1L, xmin = 0, xmax = 1.5, by = 0.01),
     palette = "topo.colors", lwd = 1, rev = TRUE)
# custom color ramp
plot(landscape(pd, degree = 2L, xmin = 0, xmax = 1.5, by = 0.01),
     palette = c("red", "green", "blue"), lwd = 1)
par(mfrow = c(1L, 1L), mar = c(5.1, 4.1, 4.1, 2.1))
}
