% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt-package.R
\docType{package}
\name{plt-package}
\alias{plt-package}
\alias{plt}
\title{'plt' package}
\description{
An Rcpp interface to Dłotko's Persistence Landscapes Toolbox
}
\details{
This package provides an R interface to the Persistence Landscapes
Toolbox by Paweł Dłotko and Peter Bubenik. It is adapted from the
'tdatools' package prepared by Jose Bouza.
}
\author{
\strong{Maintainer}: Jason Cory Brunson \email{cornelioid@gmail.com} (\href{https://orcid.org/0000-0003-3126-9494}{ORCID})

Authors:
\itemize{
  \item Jose Bouza \email{josejbouza@gmail.com} [programmer, copyright holder]
  \item Paweł Dłotko [programmer, copyright holder]
}

Other contributors:
\itemize{
  \item Ryan Elkind \email{ryanelkind@gmail.com} [contributor]
  \item Peter Bubenik [contributor, laboratory director]
}

}
