% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_persistence}
\alias{get_pairs}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{phutil}{\code{\link[phutil:persistence]{as_persistence}}, \code{\link[phutil:persistence]{get_pairs}}}
}}

