% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-run.R
\name{run_testthat}
\alias{run_testthat}
\title{Automatically Test a Set of Unit Tests in a Directory}
\usage{
run_testthat(tests_path = "testthat")
}
\arguments{
\item{tests_path}{Directory where unit tests are written}
}
\value{
A \code{list} with a format that matches the external grading specification.
See the details section for more information.
}
\description{
Parses a set of files found in the \code{testthat} directory inside a docker
container.
}
\details{
TODO: Add details or link out.
\url{https://prairielearn.readthedocs.io/en/latest/externalGrading/#grading-results}
}
\section{Directory Testing}{

By default, \code{testthat} has a preference to run tests within an \emph{R} package.
This approach uncouples the need for an \emph{R} package. Details for testing
outside of an \emph{R} package can be found at:
\url{https://github.com/r-lib/testthat/issues/659}
}

