% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show,Rcpp_PersistenceLandscape-method}
\title{Print Method for Persistence Landscapes}
\usage{
\S4method{show}{Rcpp_PersistenceLandscape}(object)
}
\arguments{
\item{object}{A \link{PersistenceLandscape} object.}
}
\description{
A \code{\link[methods:show]{methods::show()}} S4 method for persistence landscape objects,
used for \code{\link[base:print]{base::print()}}.
}
\examples{

# double matrix
pl <- new(
  PersistenceLandscape,
  cbind(c(0, 0, 1, 2), c(1.5, 2.5, 3.5, 4.5)),
  TRUE,
  0, 10, .01
)
pl$isExact()
pl$xMin()
pl$xMax()
pl$xBy()
pl$getInternal()
print(pl)
summary(pl)

dim(pl$discretize()$getInternal())
length(as.vector(pl))
as.data.frame(pl)

pl2 <- pl$scale(2)
pl2$getInternal()
pl3 <- pl$add(pl2)
pl3$getInternal()

# integer matrix
pl <- new(
  PersistenceLandscape,
  cbind(c(0L, 0L, 1L, 2L), c(2L, 3L, 4L, 5L)),
  TRUE,
  0, 10, .01
)
pl$getInternal()
print(pl)

# discrete landscape
pl <- new(
  PersistenceLandscape,
  cbind(c(0, 0, 1, 2), c(1.5, 2.5, 3.5, 4.5)),
  FALSE,
  0, 5, .5
)
pl$xBy()
pl$getInternal()
print(pl)

dim(pl$discretize()$getInternal())
length(as.vector(pl))
dim(as.matrix(pl))
dim(as.matrix(pl, by_level = TRUE))
as.data.frame(pl)

pl2 <- pl$scale(2)
pl2$isExact()
pl2$xBy()
pl2$getInternal()
show(pl2)

(pl3 <- pl$add(pl2))
pl3$getInternal()
pl3$xBy()
}
\seealso{
\linkS4class{Rcpp_PersistenceLandscape} for the exported C++ class and
\code{\link[=pl_new]{pl_new()}} for the R wrapper.
}
