% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-metadata.R
\name{test_info}
\alias{test_info}
\title{Retrieve question totals and test name}
\usage{
test_info(test_data)
}
\arguments{
\item{test_data}{A \code{data.frame} containing the test information.}
}
\value{
Modified \code{data.frame} with:
\itemize{
\item \code{max_points}: Total points available for the question.
\item \code{earned_points}: Amount of points student earned.
\item \code{title}: Name of the test
}
}
\description{
Extracts from the test data the point value and title of the test.
}
\details{
To facilitate this, we assume the format for the question is given as...

NumericPoints: title

We extract out from the test name both components.
}
